Implement a better file structure
