+++ app.js
@@ -1,6 +1,7 @@
-const fs = require('fs');
 const express = require('express');
 const morgan = require('morgan');
+const tourRoutes = require('./routes/tourRoutes');
+const userRoutes = require('./routes/userRoutes');
 
 const app = express();
 app.use(express.json());
@@ -12,110 +13,8 @@ app.use((req, res, next) => {
   next();
 });
 
-const tours = JSON.parse(
-  fs.readFileSync(`${__dirname}/dev-data/data/tours-simple.json`)
-);
-
-const createTour = (req, res) => {
-  const newId = tours[tours.length - 1].id + 1;
-  const newTour = Object.assign({ id: newId }, req.body);
-
-  tours.push(newTour);
-  fs.writeFile(
-    `${__dirname}/dev-data/data/tours-simple.json`,
-    JSON.stringify(tours),
-    (err) => {
-      res.status(201).json({ status: 'success', data: { tours: newTour } });
-    }
-  );
-};
-
-const getAllTours = (req, res) => {
-  res.status(200).json({ message: 'success', result: tours.length, tours });
-};
-
-const getTour = (req, res) => {
-  const id = req.params.id * 1;
-
-  const tour = tours.find((el) => el.id == id);
-
-  if (!tour) {
-    res.status(404).json({ status: 'fail', message: 'Invalid ID' });
-  }
-  res.status(200).json({
-    stutas: 'success',
-    data: {
-      tour,
-    },
-  });
-};
-
-const updateTour = (req, res) => {
-  if (req.params.id > tours.length) {
-    res.status(404).json({ status: 'fail', message: 'Invalid ID' });
-  }
-  res
-    .status(200)
-    .json({ status: 'success', data: '<Tour Updated successfully' });
-};
-
-const deleteTour = (req, res) => {
-  if (req.params.id > tours.length) {
-    res.status(404).json({ status: 'fail', message: 'Invalid ID' });
-  }
-  res.status(204).json({ status: 'success', data: null });
-};
-
-const getAllUsers = (req, res) => {
-  res.status(500).json({
-    status: 'error',
-    message: 'This route is not been defined',
-  });
-};
-
-const createUser = (req, res) => {
-  res.status(500).json({
-    status: 'error',
-    message: 'This route is not been defined',
-  });
-};
-
-const getUser = (req, res) => {
-  res.status(500).json({
-    status: 'error',
-    message: 'This route is not been defined',
-  });
-};
-
-const updateUser = (req, res) => {
-  res.status(500).json({
-    status: 'error',
-    message: 'This route is not been defined',
-  });
-};
-
-const deleteUser = (req, res) => {
-  res.status(500).json({
-    status: 'error',
-    message: 'This route is not been defined',
-  });
-};
-
 // ROUTES
-const tourRoutes = express.Router();
-const userRoutes = express.Router();
-
-tourRoutes.route('/').get(getAllTours).post(createTour);
-tourRoutes.route('/:id').get(getTour).patch(updateTour).delete(deleteTour);
-
-userRoutes.route('/').get(getAllUsers).post(createUser);
-userRoutes.route('/:id').get(getUser).patch(updateUser).delete(deleteUser);
-
 app.use('/api/v1/tours', tourRoutes);
 app.use('/api/v1/users', userRoutes);
 
-// SERVER START
-const port = '3000';
-app.listen(port, () => {
-  console.log(`Server is running on port ${port}...`);
-});
+module.exports = app;
