+++ package.json
@@ -4,7 +4,7 @@
   "description": "This is a study project",
   "main": "app.js",
   "scripts": {
-    "start": "nodemon app.js",
+    "start": "nodemon server.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "Mahdi Noori",
